// client/src/app/shared/chat/ESocketMessage.ts
export enum ESocketMessage {
    TRY_CREATE_CHANNEL = 'tryCreateChannel',
    CREATED_CHANNEL = 'createdChannel',
    LIST_CHANNELS = 'listChannels',
    TRY_UPDATE_CHANNEL = 'tryUpdateChannel',
    UPDATED_CHANNEL = 'updatedChannel',
    TRY_DELETE_CHANNEL = 'tryDeleteChannel',
    DELETED_CHANNEL = 'deletedChannel',
    TRY_JOIN_CHANNEL = 'tryJoinChannel',
    JOINED_TO_CHANNEL = 'joinedToChannel',
    MESSAGE = 'message',
    TRY_LEAVE_CHANNEL = 'tryLeaveChannel',
    LEFT_CHANNEL = 'leftChannel',
    TRY_INVITE_TO_CHANNEL = 'tryInviteToChannel',
    INVITED_TO_CHANNEL = 'invitedToChannel',
    TRY_MUTE_FROM_CHANNEL = 'tryMuteFromChannel',
    MUTED_FROM_CHANNEL = 'mutedFromChannel',
    TRY_KICK_FROM_CHANNEL = 'tryKickFromChannel',
    KICKED_FROM_CHANNEL = 'kickedFromChannel',
    TRY_BAN_FROM_CHANNEL = 'tryBanFromChannel',
    BANNED_FROM_CHANNEL = 'bannedFromChannel',
      TRY_ADD_ADMIN = 'tryAddAdmin',
    ADDED_ADMIN = 'addedAdmin',
    TRY_REMOVE_ADMIN = 'tryRemoveAdmin',
    REMOVED_ADMIN = 'removedAdmin',
  }